Fix parsing empty solution metadata
If solution metadata value did not contain any white spaces
after the column, the next metadata field value was read.

The problem was caused when solution was modified with an
editor and all trailing white spaces were removed. If the
metadata field did not have any value, the parser read in
incorrect value.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>