Add --headers option to print only headers
There was a problem is a search result printed too many
contents. For example a long list of solutions was hard
to use to search one solution digest to update it. Now
it is possible to print only the headers.

The usage case would be searching specific content and
getting the digest for example for update purposes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>