Fix parsing Markdown solution with '---' string
If a solution data had a string '---' anywhere in the data
section of the solution, parsing of the solution failed.
This caused a single solution to be split into a multiple
solutions. For example in case of an update operation, it
resulted a data for solution that was only a part of the
data. Also for example parsing of the 'brief' or 'category'
attributes may have failed leaving empty or default values
into the attributes.

Now the reserved string '---' must be in the beginning of a
(multi)line string and newline or end of line must follow
immediately after the string.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>