Fix unicode exception with Python 2
There are logs that use the format() function to format a
given string to the logger. This causes problems with Python 2
if the formatted string is a unicode string. In case of Python
2, the printed string must be marked with explicit u'' or
import from futures for 'unicode_literals' must be used. When
%s formatting is used, this is not a problem.

Error:

  UnicodeEncodeError: 'ascii' codec can't encode character u'\u201d'

But since the logged strings should not be interpolated always
if the logs are not printed at all. This can be done with the
%s style of format and separating the variables with comma [2].

Because of this, the formatting was changed to %s to work with
Python 2 and 3 and to comply with Logging performance aspects
as well as the Pylint warning 'logging-format-interpolation
(W1202)'.

[1] https://python-future.org/compatible_idioms.html#unicode-text-string-literals
[2] https://stackoverflow.com/a/34634301

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>