Fix Python 2 export operation with non ASCII characters
The export operation did not work with Python 2 when exported
content had other than ASCII characters.

Also there seems to be an issue with PyPy that uses Python 2.7.13.
The problem is that only 'mode' and 'buffering' parameters from
io.open can be checked. If there are additional parameters that
are tested with 'mock.assert_called_once_with()'. This seems to
cause an error from 'funcsigs/__init__.py' that complains 'too
many keyword arguments {'encoding': 'utf-8'}'. The last parameter
can be anything.

This can be solved by setting the 'autospec=True' that creates the
replaced object with spec from the object itself. This seems to
allow testing also other parameters than the 'mode' and 'buffering'.

This is not fully understood though.

Also removed the 'create=True' and 'mock.mock_open()' that seem to
be unnecessary ('create=True') and redundant ('mock.mock_open()').

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>