Fix Fix Python 2 compatibility with file open
The 'io.open' with 'utf-8' encoding is now used in all file
operations. This allows Python 2 and 3 compatible way to read
other than ASCII characters in a safer way.

The file access must use read() instead of other methods like
readlines(). The Python 2 and 3 compatible way with mocks is
to use only the read(). E.g. the readlines() mock has been
improved to Python 3. It could be possible to mock readlines()
as well, but this would require manual setting of the 'mock
return_value.readlines.return_value' (or something similar).

There seems to be also differences when mocking 'with open'
or 'with io.open'. For example multiple function calls to
'io.open' is not seen in a same way in both cases. See the
case 'test_cli_import_solution_029'.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>