Fix Docker test example command
There was an error in the example command to start the
Docker container. The --server-host option must not be
used when starting the container. This configures the
server running port which is not visible for the snippy
Docker container health check function. The environment
variable 'SNIPPY_SERVER_HOST' must be used to define
the IP and port for the server. This makes the setting
visible also for the health check function that allows
to move the container in the healthy state.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>