Add option to define pip cache for make targets
The main purpose was to be able to prevent using pip install
cache to force download of the third party package. There
may be some cases mainly with Python 3.4 where the installer
may break if cached version is used. This is complicated and
it is not fully sure if this is needed. Therefore the default
value is empty which falls to the default behaviour.

The problem was that at some point of time when working with
the Python 3.4, the cached module version was used which was
not compatibly with the Python 3.4. These may have been now
fixed since the module versions were massaged to work with
all supported Python versions.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>