Fix Python development installer instructions
There were faults with the latest virtual environment
modules. The virtual environment installation was
broken since it was not able to test the active
virtual environment correctly. The test should be
now more reliable because it uses the venv environment
variable to test if the specific environment is
active.

Added also brief instructions how to compile Python
3.8 for Fedora 28. The reason was that there are
packages needed to get the sqlite and pip to work
after compilation.

Added optional note to update pip itself before
installing the virtual environment packages. Use
this at your own risk because the command udpates
the global pip.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>