Edit Python module versions for development
The development packages were bumped for all supported
Python versions.

There are special cases to make the Python 3.4 tow work.
The tox requires specific version from importlib_metadata
for Python 3.4. Also the openapi2jsonschema requires a
specific colorama version in Python 3.4.The requirement
for tox is bit nasty since it pollutes the 'requires'
that is the official install requirement. But the
specific version lock is only in Python 3.4 so this
is considered ok.

Also with Python 3.4, it is not possible to update the
PyYAML and colorama to latest versions because dependency
issues (and possible issues in these modules) in tox
needed for Python 3.4.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>