Fix Travis CI for Windows
The Windows CI installer was vague on Python version.
Because of this, the latest Python 3.8 was installed
and the PATH was still pointing to Python 3.7.

Now there is the explicit version to use Python 3.7.5
in the installer. This causes also the PATH to be set
as '/c/Python' which is generic [1] and does not need
to be updated when the Python version changes.

[1] https://travis-ci.community/t/windows-python-pip-module-not-found/5480/8

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>