Fix Travis CI for Windows
The previous fix for Windows Travis CI failed. The path
seems to be always Pyhon version specific regardless if
'python' or 'python3' was used as installer target.

The choco installer failed with version option. The make
was tried to be installed with the python version. Now
there are separate install commands for make and python.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>