Fix documentation build in Read the Docs
The Read the Docs build tried to use the requirements.txt
with Python 2. The requirements.txt was defined for Python 3
and this failed.

Changed the Read the Docs configuration to use version 2
and updated the configuration Yaml accordingly. The Python
version is fixed and the path to Sphinx configuration file
is fixed. This tries to  provide more maintainable and
readable configuration file.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>