Add platforms metadata
There is now metadata in the setup.py file that tells which
platforms are supported. This is just a metadata to users
of this tool. The values are exactly the same as in the PyPI
Python classifiers for "Operating System" [2]. Full classifier
is not quoted, only the last part of each classifier.

[1] https://stackoverflow.com/a/34994245
[2] https://pypi.org/classifiers/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>