Test searching content with a multiword string
It was actually possible to search for example snippets with
a string that contains multiple words. The use case could be
for example like searching only content with "docker ps".

This can be achieved with search string "docker.ps" where the
dot matches to any characters as in a regexp. The original
thought was to be able to match these kind of string with an
example like "docker ps". How ever, existing implementation
already split the search string into two keywords in this
example. The dot is also considered better because it can
match also other whitespace characters like tabulators.

There was just a test case added to verify that this feature
remains functional.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>