Edit commented Sphinx path manipulation
For local environment and for the automated document build
in Read the Docs there seem to be no reason to add the
'sys.path.insert' for the Sphinx build as of now. There was
a situation that required locally setting the path. But
this started to work again.

Previously the path was '.' which did not work and it had
to be changed to '../../../snippy'. But since the 'make docs'
is working locally, this is not taken into use. Just that
the change in comments is stored for further investigation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>