Fix release documentation
The release documentation was updated when new release is
being build.

Removed a comment from development notes that was related
to installing virtual environments with no-cache-dir option.
This option was added to cope with unpinned Python modules
in the setup.py file due to Python 3.4 installation. The
comment was half a sentence and the case was not recalled
exactly.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>