Fix Python 3.4 CI build
The twine upgrade failed in case of Python 3.4. Upgrade of
the twine is added into a generic 'release' installation
package. This is not needed in case of CI tests, it is
just installed to install and upgrade all the tools in
the Python release tool chain.

The CI and tests use the latest versions from setuptools
which is part of the same release target that installs also
the twine.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>