Add support to import multiple files (WIP)
The use case is that there is a separate repository that
stores personal snippets. It is nice to be able to import
multiple files with one command. Therefore example like
below will be supported:

  $ snippy import -f *.md

Currently the existing functionality is modified to allow
globs in the filenames without breaking existing features.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>