Add Python pip upgrade strategy to 'eager'
The --upgrade-strategy option is now set by default to the
'eager'. The --upgrade option does not enforce upgrade to
the latest version if there is already a version from a
Python module that satisfies the requirement. For example
pyyaml is not pinned to make the tool to install with wide
variety of mandatory modules. For development, the latest
is preferred.

This can be controlled with Makefile variable PIP_UPGRADEi.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>