Fix incorrect error in exporting defaults
If user exported multiple content categories with the
'--default' option and some category did not have any
content, there was a error message even though some
categories were exported correctly.

It is now considered that if one category export is
successfully done when exporting with the '--default'
option, the operation is successful. If there are no
exports at all, then the same error message is set as
before.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>