Fix missing link symbol '>' from terminal
If content did not have links and it was printed on terminal,
the link symbol was not printed. It is considered that the
target is constant/static kind of terminal output. For this,
the symbols are always printed even though the data may not
be always there.

Reference content has the link attribute as a mandatory data
and thus this problem is not related to reference content.
How ever, the code to print the reference content to terminal
has same adjustment than rest of the content types. This
tries to keep the terminal output logic similar for all
content types and tries to follow logic where "styling layer
does not care about content differences".

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>