Change TODO symbol
The TODO symbol was a unicode character u'2713'. This was
changed to exclamation mark for few reasons. The unicode
character gave false impression that the TODO task was
done. Also unicode characters may be problematic in some
older terminals and the width of the textual string is not
self evident when there are unicode characters.

The exclamation mark gives an idea that "to this now!" and
it is considered to be better suited for todo tasks.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>