Fix updating todo content in Markdown format
When user first created a todo content and then updated it
in Markdown format, the data was not parsed correctly. This
was easily visible when no changes were made to the content.
Previously the data was changed and for example the digest
checksum was updated due to incorrect data parsing.

The problem was that the optional time part in the todo
was not read correctly when the todo item checksum was
added. The Checksum is not present when user adds a todo
content for the verify first time. But the tool calculates
a checksum for each todo item to directly access a specific
item. This checksum caused a parsing error when a todo
content was updated.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>